requirements.txt
whitenoise==3.3.0
psycopg2==2.7.1
pillow==4.1.0
mysqlclient==1.3.10
gunicorn==19.7.1
django==1.10.6
django-widget-tweaks==1.4.1
django-upload-avatar==0.1.0
django-registration==2.2
django-bootstrap3==8.2.1
django-appconf==1.0.2
dj-database-url==0.4.2
